feat: merge main and solve conflicts from the PR 2
